Merge Link and AnnotLink code
Annotations now belong to the Page and are created only once on demand.
Annots are now ref counted and Links is a list of AnnotLink objects,
Link object has been removed. The AnnotLink API is mostly the same than
Link and frontends APIs are not affected.
Qt4 changes made by Pino Toscano.