Fix pdftotext -htmlmeta to correctly output U+2019 in PDF metadata
In PDF documents, right single quotation mark (U+2019) may be encoded as
0x90 because PDFDocEncoding uses some of the reserved characters in
ISO 8859-1. However, pdftotext -htmlmeta assumes that characters are either
UCS-2 or ISO 8859-1. Thus when a right single quotation mark is encoded as
0x90, it is output as unicode 0x90 (which is a control character). pdfinfo
does the right thing by first converting from PDFDocEncoding to Unicode
with pdfDocEncoding[], before encoding it in the desired character set.
This patch applies the same logic to pdftotext. pdftohtml is broken in the
same way, but this patch does not attempt to fix it.

Bug 37900