xpdf303: Parse ActualText in Gfx instead of output devices
Remove beginMarkedContent and endMarkedcontent and add beginActualText
and endActualText. ActualText is parsed in Gfx, that already handles the
marked content stack, so that text output dev doesn't need to handle it
too. The text string is passed to beginActualText(). This change is not
an exact merge of xpdf code, I've tried to keep our implementation.