regtest: Add --pretty-diff option to create-report command
It includes a pretty-diff link to make a bit easier to check differences
in test results. For images it creates a html with javascript that
toggles actual and expected image, using the same code than WebKit
layout test results. For text files it uses HtmlDiff class from python
difflib. It's an option disabled by default because pretty diff for text
files is very slow.