Improve CJK suport in PSOutputDev
1. Adapt bug fix for bug 11413 to the postscript device
2. A small bug fix when locateFont doesn't find a suitable font and
returns a null pointer

to 1.:
Adapting the implementation of the bug fix for splash and cairo to the
postscript device was quite easy. But my first proofs of the output with
ghostscript 8.71 shows some regressions where the CJK chars have a
smaller height than the default square of the font. But the "48" in the
output of bug-poppler11413.pdf which is set in a "normal" font but
rotated was at the right position. Then I stepped to ghostscript 9.04,
and now the CJK chars were shown correctly, but the 48 was positioned
wrong. But because of these different tests I think that it is still a
problem in ghostscript when using a mix of CJK fonts and "normal" fonts.
BTW, also Acrobat X distiller has problems with the position of the "48"!

to 2.:
On my first tests with PDF which uses non embedded CJK fonts on Windows
I got crashes. Reason for it was that GlobalParamWin returns Helvetica,
which is not a CID font, but locateFont accepts here only CID fonts and
therefore returns a NULL pointer. I first fixed that and then decided to
return as default MS Mincho if a CID font is expected.