CJK substitute implementation on WIndows platforms.
When You install ghostscript on WIndows You're able to switch on CJK
support. This will create a  cidfmap file in the gs-lib directory. The
ps file which creates it (mkcidfm.ps) runs over the windows font
directory and tries to create a suitable substitution table for missing
CJK fonts. The cidfmap file is more or less PDF like, so it's quite easy
to parse it with our parser and create a substitution table in
GlobalParamsWin and use that table. But I expect it in the poppler data
dir instead of searching for ghostscript installation. If it is not
there, it always returns the default CID font of point 2.
You can either copy it from the gs lib directory or create it with the
ghostscript tool calling

gswin32c -q -dBATCH -sFONTDIR=<windows font directory>
-sCIDFMAP=<popper data dir>/cidfmap mkcidfm.ps