Some regression fixes/improvements
I just finished the patch for these regressions, they had differents reasons
1. In CharCodeToUnicode::mapToUnicode the identity support was missing
2. The new algorithms for axial and radial shading caused problems in
cairo. I revert these changes but removed the examination of hidden
content (this is already done in the calling function)
3. The examination of optional hidden content in showing text was wrong:
of course hidden text should not be shown, but text parameters like text
position in the state must be changed!
4. Searching and finding fonts especially with base14 fonts should be
more exact than just looking at the base14 name (i.e. fixed width and so
on) when using fontconfig. I implement that to find the best font
fitting to the needs.