Workaround Windows problem with libjpeg
In libjpeg6b, there is a type definition conflict issue in jmorecfg.h.
jmorecfg.h typedefs INT16 as short, INT32 as long. The latter conflict
with basetsd.h of Microsoft Windows SDK or MinGW that typedefs INT32
as int. basetsd.h is included by windows.h, so, if the users work with
ANSI C features only, the conflict does not appear, it would be the
background why such fundamental error has been preserved.
Anyway, jmorecfg.h designer was aware about the conflict with X11 headers
(Xmd.h defines INT16, INT32, etc), so if we pretend as if Xmd.h were
included (by defining XMD_H), this issue can be bypassed. Attached patch
checks the conflict, then, if there is a conflict, tries XMD_H trick, and
uses it if it works. If jmorecfg.h or basetsd.h was manually fixed to
prevent this issue, this trick is not used.