Add quotes
this is quite small fix. In poppler's configure, basically the shell
variables are not enclosed, and not quoted, aslike
        if test x$use_glib = x; then
This syntax make the test command confused when use_glib is multi-word
value, as, "no (required cairo output)". In such case, test cannot know
how to evaluate the token "no" "(required" "cairo", so some "syntax error"
messages are given. Thus, quoting of the variable would be safer.