Do not append "-ansi" to CXXFLAG, if "-std=XXX" is already specified.
SplashOutputDev.cc uses isfinite() function (defined by math.h).
isfinite() was standardized in C99, and imported to C++0x in later.

In QNX header file system, isfinite() is disabled by default,
the definition of __STDC_VERSION__ > 199901 is required to enable it.
In the case of GCC, "-std=c99" (for C) or "-std=gnu++0x" (for C++)
is expected.

But, current configure of poppler appends "-ansi" flag for CXXFLAGS,
if the compiler is known to be GNU. "-ansi" is C89 or C++98, so isfinite()
is unavailable. I propose a patch NOT to "-ansi" flag, if CXXFLAGS
includes "-std=XXX" already.