text: increase the tolerance for overlapping glyphs
TextOutputDev will start a new line when encountering consecutive
glyphs with overlapping bounding boxes. This can occur when drawing
diacritics with a separate glyph. In this case, due to the diacritic
having a different baseline, the lines may be output in the wrong
order.

This patch increases the tolerance for overlapping bounding boxes to
prevent diacritics from splitting lines.