Splash: Avoid bogus memory error for tilingPattern
just return gFalse if the resulting
tiling pattern bitmap reaches a memory limit and so the Gfx implemention of
tiling patterns will be used. I think that this is an acceptable approach also
concerning performance, because it would take also a while to paint such a huge
bitmap and then draw it to splash.