FormFieldChoice ctor: Don't convert "human-readable" option names to unicode
Despite that comment, they're not meant to be read by humans only, but they
are also used as option identifiers.

This patch stops poppler from forcing them to be unicode. Instead,
they now stay the same encoding as their corresponding /Opt entry.

This fixes poppler not being able to recognize selected entries
in documents produced by poppler itself: previously, the /V value was
always written in Unicode encoding, and therefore it was very often not
binary-equal to the corresponding /Opt entry.
Now the /V value is always binary-equal to the corresponding /Opt entry.