Replaced forward declaration of enum CryptAlgorithm with a proper #include
Forward-declaring enums is illegal. It hasn't caused problems so far
because XRef.h includes Object.h, which in turn includes Stream.h, where
the enum is defined.
Therefore, enum is alreadly defined when the compiler reaches that line.

To avoid future issues, I've replaced it with a proper #include "Stream.h"
(which expands to nothing as Stream.h has already been included at that
point).

I've also added a #include in XRef.h because it references enum
CryptAlgorithm too. Also in this case, it expands to nothing.