cairo: make drawImage work with images > 32767 in width/height
Cairo images are limited to 32767 in width and height due to the
16.16 format used by pixman. Make drawImage work with large images
by scaling down the image before a cairo image is created.

CairoRescaleBox.cc has been turned into a class with a virtual
function to get the next row of the source image. This allows the
rescale code to access the source data one row at a time to avoid
needing to allocate an image the size of the source image.

A RescaleDrawImage class derived from CairoRescaleBox has been
written to create the cairo source image to be used by drawImage. The
code from drawImage that created the cairo source image has been moved
into RescaleDrawImage::getSourceImage and RescaleDrawImage::getRow.

Bug 56858