Repair pdfunite
Sorry, when I implemented the support encrypted pdf files in pdfseparate I
missed that writePageObjects of course is also used in pdfunite for combining
pages, and even more that encrypted files are still not supported by pdfunite,
I removed the numoffset from writing the objects itself. Therefore there are
still all objects in the combined pdf file, but the references missing the
numoffset and therefore were no more reachable.
The patch repairs it.

Bug #58569