Parser: return error if stream encountered when allowStreams = false
Opening a PDF file where the first object is a stream prints a
"Command token too long" error message. This is caused by the
Linearization check in Linearization::Linearization reading objects
with allowStreams = false. The Parser ignores the "stream" token and
tries reading the next token which is usually binary data. Setting
allowStreams to true will not work since the stream length is often an
indirect object and at this point the XRef has not been created.

Fix this by making Parser return an error object if the "stream" token
is encountered when allowStreams is false.

Bug 58966