Fix for complete rewrites in repaired files
Part 2 of bug #62985
if You save a PDF with defect xref offsets, the

readXRefUntil(-1 /* read all xref sections */, &xrefStreamObjNums)

in XRef::scanSpecialFlags() will destroy the already reconstructed entries table, but this means that any modification which the user did in the meantime get lost. This can be tested i.e. with bug168518.pdf.