cairo: Always use the trasnfer function if present in setSoftMask
"If the subtype is Alpha, the transparency group XObject G shall be
evaluated to compute a group alpha only. The colours of the constituent
objects shall be ignored and the colour compositing computations shall
not be performed. The transfer function TR shall then be applied to the
computed group alpha to produce the mask values."

We were using the transfer function only for luminosity soft masks.

https://bugs.freedesktop.org/show_bug.cgi?id=63587