glib: Use TextPage::getSelectionWords to build text layout and attributes
This way we can make sure that the list of words used in
poppler_page_get_text_layout and poppler_page_get_text_attributes is the
same that the one used in poppler_page_get_text. This fixes the mismatch
between the number of characters in the text returned by
poppler_page_get_text and the number of characters returned by
poppler_page_get_text_layout in some documents.