Make some pdftops conversions *much* faster
For example: http://ev.kde.org/resources/expense_report.pdf

Without this patch it seems "infinite", which this patch it's a few seconds

The change: Instead of just remembering in xobjStack the set of XObjects (and Patterns,
the variable name was 'wrong') we are currently setting up (i.e. the current chain), we
remember all of them.

This has passed the pdf->ps regression test without a single issue