TextOutputDev: Honor the selection rectangle passed to TextPage::getSelectionWords()
Make TextPage::getSelectionWords() return a list of TextWordSelection
instead of a list of TextWord so that it's possible to know which
characters of the word are inside the given selection rectangle.
Adapt the glib frontend to the new API and use the selection bounds
instead of the whole word to build the list of characters in
poppler_page_get_text_layout() and poppler_page_get_text_attributes(),
which ensures the number of glyphs returned is in sync with the number
of characters returned by poppler_page_get_text().

https://bugs.freedesktop.org/show_bug.cgi?id=71160