regtest: Read test results in chunks to get the md5 digest
Some backends can generate huge results, like huge postscript files that
we don't want to load in memory to get the md5. So, instead of creating
thr md5 object with the entire file, we feed it with chunks of data
using the update method. This improves a lot the memory consumption and
performance as well.