regtest: Limit the stderr files to ~1MB
Some tests send a lot of information to stderr, usually due to parsing
errors in buggy documents. More than 1MB of stderr output is diffcult to
hanlde and in most cases it's redundant with a lot of duplicated messages.
This patch reduced the size of the refs dir for the complete test suite
by 1GB.