glib: Remove unneeded POPPLER_STRUCTURE_ELEMENT_UNKNOWN
The enum value POPPLER_STRUCTURE_ELEMENT_UNKNOWN because Poppler does
not add invalid StructElements ("invalid" being nodes of type
StructElement::Unknown) to the Tagged-PDF structure tree. That means
that poppler-glib does not need to expose it in the API. An assertion
is left to aid in finding issues when using debug builds.

https://bugs.freedesktop.org/show_bug.cgi?id=75541