cairo: Implement function shading using mesh gradients
Gfx draws function shadings by subdividing the shading until the
colors are the same or the maximum subdivision is reached then fills
each cell with the color of the mid point of the cell. The solid
colors can result in a pixelated appearance.

This patch implements a cairo specific version of the function shading
that uses mesh gradients to draw each cell. By setting the corner of
each patch to the shading color at that point, the mesh gradient will
interpolate the colors resulting in a smooth appearance.

Bug 88394