Fix regression in pdftops parameter passing
Recent changes to the command line processing moved some options from GlobalParams to PSOutputDev.
The options used to be set by calling setters in GlobalParams before creating the PSOutputDev.
Now the options are set by calling setters in PSOutputDev after it is created.
The problem is that PSOutputDev() calls init() which uses options that now can not be set until later after PSOutputDev() returns.
These patches split some of the code of init() into a new postInit() that is automatically called later after the pdftops main program has had a chance to set the command line options.
When I was looking through the code, I also fixed a misspelling of sanitizedTitle as sanitizedTile.

Bug #89827