Add KeepAlphaChannel render flag to Qt frontends
Adds a new render flag which will indicate that the image return by
Poppler::Page::renderToImage will not be opaque and alpha blended with
the paper colour, but retain its actually background transparency.

This improves performance in synthentic rendering benchmarks by almost
five percent and the additional alpha blending that is then done by the
consuming application is often a hardware-accellerated operation.