cpp: Fix utf8/utf16 conversion
The old code assumed that ustring::size() would return the number of
bytes in ustring, but it really returns the number of characters. Since
ustring is a basic_string<unsigned short>, these two values differ (by a
factor of two).

This needs to be considered when using iconv, since it operates on byte
counts, not character counts.

Bug #91644