regtest: Add support for testing password protected documents
Similar to how skipped file works, you can pass a passwords file from
the command line or add a Passwords file to your docs directory. This
file should be a python file containing a "passwords" dictionary where
the key is the test document and the value is the password required to
open that document.