cairo: use shape mask with soft mask
Clear target with the shape mask whether the soft mask is set or not.
Propagate the shape up to any higher level groups and destroy the
shape pattern when done.

Fix a memory leak by removing a call to 'cairo_reference (cairo_shape)'.
We already keep track of the lifetime of cairo_shape using
knockoutCount.

bug 91931