introduced hex string as a new Object type and used it for file identifier
File identifiers are usually written as hex strings (and this is how the PDF
reference presents them in an example). Until now, poppler was reading hex
strings properly, but was forgeting about the fact that a given string is a hex
string, so e.g. file identifier was first read as a hex string and then printed as
an ordinary string and thanks to that what was printed was actually junk. This
commit fixes that.