pdfimages: don't fail listing if inline image data contains 'EI'
Normally when listing images we don't read the image data. But for
inline images we should read the image data to advance the stream
position to the end of the image data. If we don't advance the stream
position and the image data happens to contain 'EI', Gfx will resume
reading the content stream from the middle of the image data.

Bug 100737