configure.ac: fix --disable-FEATURE actually enabling the feature
Forwarded: https://bugs.freedesktop.org/show_bug.cgi?id=101818

A frequently seen antipattern is to use
AC_ARG_ENABLE(feature, help, action-if-given, action-if-not-given) as
AC_ARG_ENABLE(feature, help, action-if-enabled, action-if-disabled).
However, action-if-given is also evaluated for --disable-FEATURE (with
enableval=no), which results in --disable-FEATURE and --enable-FEATURE
doing the same in this case.

Rewrite the single-precision, fixedpoint and cmyk arguments accordingly
so the user is not confused if they explicitely want to disable those
options.

Signed-off-by: Roland Hieber <r.hieber@pengutronix.de>