qt5: ArthurOutputDev: Properly set the QPainter transformation
Previously, the code did not touch the QPainter world transformation
matrix.  Rather, the code contained ad hoc coordinate transformations
scattered throughout the code.  This patch does a cleanup: the QPainter
transformation matrix is set properly, and the hand-coded
transformations are removed.  This should not affect the rendering
output, but it makes the current (and future!) code simpler to
read and write.