cmake: restructure the warnings into a "default" and "extra" group
Most of the previous warnings are now in default. Extra warnings that
should be fixed but currently result in a lot of warnings are in the
extra group.

The old no/yes/kde COMPILE_WARNINGS option has been replaced with the
boolean option EXTRA_WARN (default off) to enable the extra warnings.