Error out on save if file has changed since we opened it
In poppler we keep the fd of the file open so the XRef+FileStream can
locate objects. This is good since we save lots of memory for not having
everything on memory all the time, but that means that when we want to
save we need the file to be exactly the same as it was when we created
the XRef otherwise we're going to be reading from the wrong part of the
"new" file.

Bug #103793