ArthurOutputDev: Replace the QPainter by a stack of QPainters
This patch lays some groundwork for the support of transparency
groups.  Transparency groups temporarily create new painters.
These get painted upon, and then the resulting new painting
will be drawn at once onto the original PaintDevice.  To implement
this, we need a stack of painters rather than a single one.
The first painter on the stack is the original one.  Opening a
transparency group pushes a new painter onto the stack, and
all drawing operations always go to the painter on the top
of the stack.